clean: create helpers for deleteModule in utils.ts for minify that func
